requirements.txt
python-telegram-bot==20.8
pyotp==2.9.0